# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "toshers",
    "version": "0.1.0",
    "description": "A curated REST + MCP library over Henry Mayhew's 1851 London Labour and the London Poor, Vol. II — the street-finders (bone-grubbers, pure-finders, sewer-hunters/toshers, mud-larks, dustmen). Every claim is line-citable against source.txt. No auth, CORS wide open.",
    "license": { "name": "Public domain (source); code MIT", "identifier": "MIT" },
    "contact": { "url": "https://github.com/skyblue-will/toshers" }
  },
  "servers": [
    { "url": "https://toshers.vercel.app", "description": "Production" }
  ],
...
  "paths": {
...
    "/api/chapters/{id}": {
      "get": {
        "tags": ["chapter"],
        "summary": "One chapter",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["raw"] }, "description": "If 'raw', returns the full markdown file as text/markdown" },
          { "name": "profile", "in": "query", "schema": { "type": "string", "enum": ["minimal", "facts", "full"], "default": "full" }, "description": "'minimal' = metadata only (no body); 'facts' = frontmatter minus game_hooks + body; 'full' = everything." }
        ],
        "responses": {
          "200": {
            "description": "Chapter",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Chapter" } },
              "text/markdown": { "schema": { "type": "string" } }
            }
          },
          "404": { "description": "Chapter id not found" }
        }
      }
    },
...
    "/api/characters/{id}": {
      "get": {
        "tags": ["character"],
        "summary": "One character testimony",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["raw"] } },
          { "name": "profile", "in": "query", "schema": { "type": "string", "enum": ["minimal", "facts", "full"], "default": "full" }, "description": "'minimal' = metadata only (no body); 'facts' = frontmatter minus game_hooks + body; 'full' = everything." }
        ],
        "responses": {
          "200": {
            "description": "Character",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Character" } },
              "text/markdown": { "schema": { "type": "string" } }
            }
          },
          "404": { "description": "Character id not found" }
        }
      }
    },
...
    "/api/prices/normalize": {
      "get": {
        "tags": ["synthesis"],
        "summary": "Pre-decimal £/s/d → modern GBP on four economic bases",
        "description": "Returns real_price (CPI), labour_value (avg earnings), income_value (GDP per capita), and economic_share (GDP share). These differ by an order of magnitude — use the one appropriate to what you are comparing (consumer goods vs. wages vs. industry totals). Accept either the numeric triple or a single `literal` string (e.g. '£3 5s 6d', '6d', '3l. 5s.', '£148,000').",
        "parameters": [
          { "name": "literal", "in": "query", "schema": { "type": "string" }, "description": "Pre-decimal literal, e.g. '£3 5s 6d'. Takes precedence over pounds/shillings/pence if both are passed." },
          { "name": "pounds", "in": "query", "schema": { "type": "number", "minimum": 0 } },
          { "name": "shillings", "in": "query", "schema": { "type": "number", "minimum": 0 } },
          { "name": "pence", "in": "query", "schema": { "type": "number", "minimum": 0 } }
        ],
        "responses": {
          "200": { "description": "Normalized price", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NormalizedPrice" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Chapter": {
        "type": "object",
        "required": ["id", "url", "meta", "body"],
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string" },
          "source_lines": { "type": ["string", "null"] },
          "meta": { "type": "object", "additionalProperties": true },
          "body": { "type": "string" }
        }
      },
      "Character": {
        "type": "object",
        "required": ["id", "url", "meta", "body"],
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string" },
          "source_lines": { "type": ["string", "null"] },
          "meta": { "type": "object", "additionalProperties": true },
          "body": { "type": "string" }
        }
      },
...
      "BasisResult": {
        "type": "object",
        "required": ["key", "label", "modern_gbp_approx", "multiplier", "endpoint_year", "description", "source"],
        "properties": {
          "key": { "type": "string", "enum": ["real_price", "labour_value", "income_value", "economic_share"] },
          "label": { "type": "string" },
          "modern_gbp_approx": { "type": "number" },
          "multiplier": { "type": "number" },
          "endpoint_year": { "type": "integer" },
          "description": { "type": "string" },
          "source": { "type": "string" }
        }
      },
      "NormalizedPrice": {
        "type": "object",
        "required": ["input", "decimal_pounds_1851", "bases", "modern_gbp_approx", "basis", "basis_source", "note"],
        "properties": {
          "input": {
            "type": "object",
            "properties": {
              "pounds": { "type": "number" },
              "shillings": { "type": "number" },
              "pence": { "type": "number" }
            }
          },
          "decimal_pounds_1851": { "type": "number" },
          "bases": {
            "type": "array",
            "description": "Four parallel modern-GBP equivalents on different economic bases.",
            "items": { "$ref": "#/components/schemas/BasisResult" }
          },
          "modern_gbp_approx": { "type": "number", "description": "Back-compat alias for bases[real_price].modern_gbp_approx." },
          "basis": { "type": "string", "description": "Back-compat alias for bases[real_price].label." },
          "basis_source": { "type": "string", "description": "Back-compat alias for bases[real_price].source." },
          "note": { "type": "string" }
        }
      },
...
    }
  }
}